requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.115.6
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.0
fpdf==1.7.2
fpdf2==2.8.1
frozenlist==1.5.0
# Editable install with no version control (fwt_rankings==0.1.0)
-e c:\users\kaiun\documents\liveheats\code\fwt_rankings
gql==3.5.0
graphql-core==3.2.5
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.0.0
propcache==0.2.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.17.1
